+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: nloomans <nloomans@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/03/23 15:31:40 by nloomans       #+#    #+#                */
-/*   Updated: 2019/03/28 20:11:58 by nloomans      ########   odam.nl         */
+/*   Updated: 2019/03/28 20:19:18 by nloomans      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *b, int c, size_t len)
 {
-	unsigned long	i;
+	size_t	i;
 
 	i = 0;
 	while (i < len)
